+++ src/components/workshops/search/SearchInput.js
